tidy up test files
